# [CARBONDATA-3637] Optimize insert into flow